Address code review comments on initial commit
Primarily this change removed generated code for the type clients, listers,
and informers, and replaces those functions with dynamic versions to more
closely match what kube builder does. The main benefit was to remove a
dependency on k8s.io/code-generator, and to simplify keeping up to date
with the latest k8s release.